Implement path references in terms of bytestring expressions
This patch re-implements path references (&label which expands to the
referenced node's full path) in terms of the bytestring expression
infrastructure.

Internally this is implemented as a constant bytestring expression
containing nothing but the marker to insert the path reference later.
Logically it would be neater to have the expression itself expand directly
to the right value itself, but for now we're evaluating these expressions
before all the tree's paths have been constructed.

Signed-off-by: David Gibson <david@gibson.dropbear.id.au>