Re-implement "," in property definitions as a bytestring operator
We've already introduced an internal "join" operator which appends
bytestring expressions together.  This uses it to implement the "," syntax
in property values as expressions.

This is the last piece of property value syntax to be converted to use
the expression infrastructure, so all (non-empty) property values can now
be implemented as a single expression.  For now we still just immediately
evaluate the expression though.

Signed-off-by: David Gibson <david@gibson.dropbear.id.au>